Chore: move getClassNames into separate stylingUtils
