+++ src/templates/contact-us/LocationsSection.jsx
@@ -4,7 +4,7 @@ import { forwardRef } from "react"
 
 import editorStyles from "styles/isomer-cms/pages/Editor.module.scss"
 
-import { getClassNames } from "utils"
+import { getClassNames } from "templates/utils/stylingUtils"
 
 const LocationHours = ({ operatingHours }) => (
   <div className={getClassNames(editorStyles, ["col", "is-6"])}>
