+++ src/templates/homepage/InfobarSection.jsx
@@ -3,7 +3,7 @@ import { forwardRef } from "react"
 
 import editorStyles from "styles/isomer-cms/pages/Editor.module.scss"
 
-import { getClassNames } from "utils"
+import { getClassNames } from "templates/utils/stylingUtils"
 
 /* eslint
   react/no-array-index-key: 0
